rename files, bugfix in src/generator.ts
