+++ examples/example-fetch/fetch.ts
