+++ examples/example-ping-pong/ping_pong.ts
