+++ examples/example-ping-pong/ping_sm.cpp
