+++ examples/example-ping-pong/ping_test.cpp
