+++ package.json
@@ -3,11 +3,14 @@
   "version": "1.0.4",
   "description": "C++ code generator for Xstate State Machine",
   "main": "dist/index.js",
+  "source": "src/index.ts",
   "types": "dist/index.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "ts-node src/index.ts",
-    "copy": "find ./src \\( -name '*.h' -o -name '*.cpp' \\) -type f -exec cp {} ./dist \\;"
+    "copy": "find ./src \\( -name '*.h' -o -name '*.cpp' \\) -type f -exec cp {} ./dist \\;",
+    "build": "tsc",
+    "prepare": "npm run build"
   },
   "repository": {
     "type": "git",
