+++ tsconfig.json
@@ -5,7 +5,7 @@
         "lib": ["es5", "es6", "dom"],
         "moduleResolution": "node",
         "declaration": true,
-        "outDir": "./dist",
+        "outDir": "./dist"
     },
     "include": [
         "src/**/*"
